app/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
redis==7.2.1
rq==2.7.0
gunicorn==25.1.0
gevent==25.9.1
greenlet==3.3.2
python-dotenv==1.2.2
Werkzeug==3.1.6
waitress==3.0.2
mysql-connector-python==9.6.0
Jinja2==3.1.6
MarkupSafe==3.0.3
click==8.3.1
itsdangerous==2.2.0
blinker==1.9.0
